HealthAnalyzer/backend/requirements.txt
flask
python-dotenv
requests
supabase
python-jose
gunicorn
flask-cors
pytest